+++ changelog.md
@@ -16,9 +16,10 @@
   * add sass
   * add material ui
   * functioning timer
-  * auto session/break swap
   * proper "XX:XX" time format on display
   * placeholder alarm
 
 ## todo
  * style (SCSS & material ui)
+ * remove ispause state instead alter basis directly
+ * auto session/break swap